chore: solve issues when add dependencies in HMR
